# docs/3차/db_api/wooriai_phase3_openapi_v0_3.yaml
openapi: 3.1.0
info:
  title: 우리아이 API
  version: 0.3.0
  description: '우리아이 MVP 개발 고정용 REST API 명세. Base URL: /api/v1'
servers:
- url: https://api.wooriai.app/api/v1
  description: Production placeholder
- url: http://localhost:3000/api/v1
  description: Local
security:
- BearerAuth: []
...
paths:
...
  /children/{childId}/expenses:
    get:
      tags:
      - Expenses
      operationId: listExpenses
      summary: 월별 지출 목록
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  expenses:
                    type: array
                    items:
                      $ref: '#/components/schemas/Expense'
                  totalAmountKrw:
                    type: integer
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      parameters:
      - $ref: '#/components/parameters/ChildId'
      - $ref: '#/components/parameters/YearMonth'
    post:
      tags:
      - Expenses
      operationId: createExpense
      summary: 지출 생성
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      parameters:
      - $ref: '#/components/parameters/ChildId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseRequest'
  /expenses/{expenseId}:
    get:
      tags:
      - Expenses
      operationId: getExpense
      summary: 지출 상세
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      parameters:
      - name: expenseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags:
      - Expenses
      operationId: updateExpense
      summary: 지출 수정
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      parameters:
      - name: expenseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                categoryId:
                  type: string
                  format: uuid
                amountKrw:
                  type: integer
                spentOn:
                  type: string
                  format: date
                itemName:
                  type: string
                memo:
                  type: string
    delete:
      tags:
      - Expenses
      operationId: deleteExpense
      summary: 지출 삭제
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      parameters:
      - name: expenseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
...
  /children/{childId}/reports/monthly:
    get:
      tags:
      - Reports
      operationId: getMonthlyReport
      summary: 월별 리포트
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportMonthly'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      parameters:
      - $ref: '#/components/parameters/ChildId'
      - $ref: '#/components/parameters/YearMonth'
  /children/{childId}/reports/cumulative:
    get:
      tags:
      - Reports
      operationId: getCumulativeReport
      summary: 누적 리포트
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalExpenseKrw:
                    type: integer
                  yearly:
                    type: array
                    items:
                      type: object
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      parameters:
      - $ref: '#/components/parameters/ChildId'
  /children/{childId}/reports/category:
    get:
      tags:
      - Reports
      operationId: getCategoryReport
      summary: 카테고리별 리포트
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items:
                      type: object
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      parameters:
      - $ref: '#/components/parameters/ChildId'
      - $ref: '#/components/parameters/YearMonth'
  /children/{childId}/items:
    get:
      tags:
      - Items
      operationId: listItems
      summary: 준비템 목록
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ItemSummary'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      parameters:
      - $ref: '#/components/parameters/ChildId'
      - name: tab
        in: query
        required: false
        schema:
          type: string
          enum:
          - now
          - soon
          - prepared
          - not_needed
  /children/{childId}/items/{itemTemplateId}:
    get:
      tags:
      - Items
      operationId: getItemDetail
      summary: 준비템 상세
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemDetail'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      parameters:
      - $ref: '#/components/parameters/ChildId'
      - name: itemTemplateId
        in: path
        required: true
        schema:
          type: string
          format: uuid
  /children/{childId}/items/{itemTemplateId}/status:
    patch:
      tags:
      - Items
      operationId: updateItemStatus
      summary: 준비템 상태 변경
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemSummary'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      parameters:
      - $ref: '#/components/parameters/ChildId'
      - name: itemTemplateId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - status
              properties:
                status:
                  type: string
                  enum:
                  - not_prepared
                  - prepared
                  - gifted
                  - not_needed
                  - interested
                expenseId:
                  type: string
                  format: uuid
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
              example: EXPENSE_AMOUNT_INVALID
            message:
              type: string
              example: 금액을 다시 확인해주세요.
            details:
              type: object
              additionalProperties: true
            requestId:
              type: string
    MoneyKRW:
      type: integer
      minimum: 1
      example: 49800
...
    Expense:
      type: object
      required:
      - id
      - childId
      - amountKrw
      - spentOn
      - itemName
      properties:
        id:
          type: string
          format: uuid
        childId:
          type: string
          format: uuid
        categoryId:
          type: string
          format: uuid
        amountKrw:
          $ref: '#/components/schemas/MoneyKRW'
        spentOn:
          type: string
          format: date
        itemName:
          type: string
          example: 기저귀
        merchant:
          type:
          - string
          - 'null'
        memo:
          type:
          - string
          - 'null'
        expenseType:
          type: string
          enum:
          - expense
          - gift
          - refund
          default: expense
        source:
          type: string
          enum:
          - manual
          - excel_import
          - purchase_followup
          - admin
    CreateExpenseRequest:
      type: object
      required:
      - categoryId
      - amountKrw
      - spentOn
      - itemName
      properties:
        categoryId:
          type: string
          format: uuid
        amountKrw:
          $ref: '#/components/schemas/MoneyKRW'
        spentOn:
          type: string
          format: date
        itemName:
          type: string
          maxLength: 120
        merchant:
          type: string
        paymentMethod:
          type: string
          enum:
          - unknown
          - cash
          - card
          - transfer
          - mobile_pay
          default: unknown
        memo:
          type: string
        linkedItemTemplateId:
          type: string
          format: uuid
...
    ItemSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        necessityLevel:
          type: string
          enum:
          - essential
          - convenience
          - optional
        status:
          type: string
          enum:
          - not_prepared
          - prepared
          - gifted
          - not_needed
          - interested
        timingLabel:
          type: string
        priceBandText:
          type: string
    ItemDetail:
      allOf:
      - $ref: '#/components/schemas/ItemSummary'
      - type: object
        properties:
          reasonText:
            type: string
          skipReasonText:
            type:
            - string
            - 'null'
          usedSecondhandOk:
            type: boolean
          safetyNote:
            type:
            - string
            - 'null'
          productLinks:
            type: array
            items:
              $ref: '#/components/schemas/ProductLink'
    ProductLink:
      type: object
      properties:
        id:
          type: string
          format: uuid
        platform:
          type: string
          enum:
          - coupang
          - naver
          - custom
        title:
          type: string
        isAffiliate:
          type: boolean
        isSponsored:
          type: boolean
        disclosureText:
          type: string
          example: 이 링크로 구매하면 우리아이가 수수료를 받을 수 있어요.
...
    ReportMonthly:
      type: object
      properties:
        childId:
          type: string
          format: uuid
        yearMonth:
          type: string
          format: date
        totalExpenseKrw:
          type: integer
        budgetAmountKrw:
          type:
          - integer
          - 'null'
        categoryTop:
          type: array
          items:
            type: object
...
  parameters:
    ChildId:
      name: childId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    YearMonth:
      name: yearMonth
      in: query
      required: false
      schema:
        type: string
        pattern: ^\d{4}-\d{2}$
        example: 2026-07